server/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
PyJWT==2.9.0
bcrypt==4.2.0
python-multipart==0.0.9
typing-extensions>=4.12.2
